le rTime est un int now

+++ fr.ensisa.hassenforder.game.empty.server/src/fr/ensisa/hassenforder/game/server/Writer.java
@@ -96,7 +96,7 @@ public class Writer extends BasicAbstractWriter {
 					Category category = p.getCategory();
 					writeCategory(category);
 					writeString(name);
-					writeLong(rTime);
+					writeInt((int)rTime);
 					writeBoolean(count > 1);
 					writeInt(count);
 					writeImage(img);
